Move tests of PaletteResultsList into PaletteResultsItem.unit.spec
